change Double to BigDecimal and added the getters and setters

+++ src/main/java/com/academy/billing/model/Billing.java
@@ -31,11 +31,11 @@ public class Billing extends BaseAuditClass{
         this.name = name;
     }
 
-    public Double getAmount() {
+    public BigDecimal getAmount() {
         return amount;
     }
 
-    public void setAmount(Double amount) {
+    public void setAmount(BigDecimal amount) {
         this.amount = amount;
     }
